+++ src/test/Global.h
@@ -24,6 +24,7 @@
 #include <unordered_set>
 #include <vector>
 
+#include "Tags.h"
 #include "TestObject.h"
 #include "../utility/Singleton.h"
 
@@ -41,6 +42,11 @@ public:
   template<typename T> T fetchData(const std::string& column);
   void runTestsAndLog();
   std::size_t num_failures() const;
+  template<typename T> bool almostEqual(T actual, T expected);
+  void abs_tol(double value);
+  void rel_tol(double value);
+  void xfail();
+  void tagRow(const std::string& row, tags tag);
 
 private:
   void addTest(std::string);