+++ src/test/Interface.cpp
@@ -1,4 +1,4 @@
-/* Copyright 2020 Ole Kliemann, Malte Kliemann
+/* Copyright 2021 Ole Kliemann, Malte Kliemann
  *
  * This file is part of DrMock.
  *
@@ -16,33 +16,46 @@
  * along with DrMock.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "test/Test.h"
-#include "mock/VerifyAllMock.h"
+#include "Interface.h"
 
-using namespace outer::inner;
+#include "SkipTest.h"
 
-DRTEST_TEST(fails)
+namespace drtest {
+
+void
+abs_tol(double value)
 {
-  VerifyAllMock foo{};
-  foo.mock.f().push().expects(1).times(1);
-  foo.mock.g().push().expects(1.23f, 4.56).times(1);
-
-  // Call only `f`.
-  foo.f(1);
-
-  // Should fail as `g` was not called as expected.
-  DRTEST_ASSERT_TEST_FAIL(DRTEST_VERIFY_MOCK(foo.mock));
+  drutility::Singleton<detail::Global>::get()->abs_tol(value);
 }
 
-DRTEST_TEST(succeeds)
+void
+rel_tol(double value)
 {
-  VerifyAllMock foo{};
-  foo.mock.f().push().expects(1).times(1);
-  foo.mock.g().push().expects(1.23f, 4.56).times(1);
-
-  // Call both methods.
-  foo.f(1);
-  foo.g(1.23, 4.56);
-
-  DRTEST_VERIFY_MOCK(foo.mock);
+  drutility::Singleton<detail::Global>::get()->rel_tol(value);
 }
+
+void
+tagRow(const std::string& row, tags tag)
+{
+  drutility::Singleton<detail::Global>::get()->tagRow(row, tag);
+}
+
+void
+skip()
+{
+  skip("");
+}
+
+void
+skip(std::string what)
+{
+  throw SkipTest{what};
+}
+
+void
+xfail()
+{
+  drutility::Singleton<detail::Global>::get()->xfail();
+}
+
+} // namespace drtest